Refactored BasicPointOps to allow sharing of smoothing code.

+++ src/main/scala/com/massivedatascience/clusterer/KMeansInitializer.scala
@@ -21,7 +21,7 @@ import org.apache.spark.rdd.RDD
 
 trait KMeansInitializer extends Serializable {
   def init(ops: BregmanPointOps, d: RDD[BregmanPoint], numClusters: Int,
-    initial: Option[Seq[IndexedSeq[BregmanCenter]]] = None, runs: Int, seed: Int): Array[Array[BregmanCenter]]
+    initial: Option[Seq[IndexedSeq[BregmanCenter]]] = None, runs: Int, seed: Long): Array[Array[BregmanCenter]]
 }
